airflow/requirements.txt
requests==2.27.1
psycopg2-binary
google-cloud-bigquery
